desktop-agent/requirements.txt
mss>=9.0.0
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
requests>=2.31.0
